Updated until "Shaming" section
Added a new exercise to the "Seguro Popular" section, changed the graphs visual aspects and labels in the sections "Seguro Popular", "Kenya Voting", and "National Health and Nutrition Survey." Moreover changed the exercise texts and added a caption to the Kenya Voting graph. 
Overall, I updated and revised three of the four sections in this tutorial. I will edit "Shaming" section later.