Updated the tutorial section
Fixed indentations, although there might be a few I have missed. Wrote the introduction, add comments to the code, changed some of the errors and problematic codes. For instance, I shortened some of the occupation names in the Gender Inequality graph and added annotation to show which value is which gender.

Existing Problems:
There might be a few indentations that are not stylistically correct.
Using of quotations is not the same throughout the tutorial, which I will work on soon.